Take migration to MariaDB as an opportunity to review data types and simplify migrations.

+++ db/migrate/20141217163230_create_events.rb
@@ -1,8 +1,8 @@
 class CreateEvents < ActiveRecord::Migration
   def change
     create_table :events do |t|
-      t.integer :system_id
-      t.integer :service_id
+      t.bigint :system_id
+      t.bigint :service_id
       t.datetime :generated
       t.datetime :stored
