+++ db/migrate/20150214223150_create_comments.rb
@@ -3,7 +3,7 @@ class CreateComments < ActiveRecord::Migration
     create_table :comments do |t|
       t.text :message
       t.integer :user_id
-      t.integer :alert_id
+      t.bigint :alert_id
 
       t.timestamps
     end
