+++ db/migrate/20150310002140_create_display_fields.rb
@@ -2,9 +2,9 @@ class CreateDisplayFields < ActiveRecord::Migration
   def change
     create_table :display_fields do |t|
       t.integer :display_id
-      t.integer :field_position
+      t.smallint :field_position
       t.string :title
-      t.integer :order
+      t.smallint :order
 
       t.timestamps
     end
