+++ db/migrate/20150310143830_add_key2_to_display.rb
@@ -1,6 +1,6 @@
 class AddKey2ToDisplay < ActiveRecord::Migration
   def change
     add_column :displays, :key2, :string
-    add_column :displays, :key2_field, :integer
+    add_column :displays, :key2_field, :smallint
   end
 end
